Patch 6.2.109 Problem:    Compiler warnings with various Amiga compilers. Solution:   Add typecast, prototypes, et al. that are also useful for other 	    systems.  (Flavio Stanchina) Files:	    src/eval.c, src/ops.c

+++ src/eval.c
@@ -9552,8 +9552,8 @@ get_return_cmd(retvar)
 	sprintf((char *)IObuff, ":return \"%s\"",
 		((VAR)retvar)->var_val.var_string);
     else
-	sprintf((char *)IObuff, ":return %d",
-		((VAR)retvar)->var_val.var_number);
+	sprintf((char *)IObuff, ":return %ld",
+		(long)(((VAR)retvar)->var_val.var_number));
     return vim_strsave(s);
 }
